refactor: move fuctions getting prerequisite to feature domain
