+++ pkg/gateway/api/api_grpc_test.go
@@ -2240,278 +2240,6 @@ func TestGetTargetFeatures(t *testing.T) {
 	}
 }
 
-var allFeaturesForPrerequisiteTest = map[string]*featureproto.Feature{
-	"featureA": {
-		Id:   "featureA",
-		Name: "featureA",
-		Prerequisites: []*featureproto.Prerequisite{
-			{
-				FeatureId: "featureE",
-			},
-			{
-				FeatureId: "featureF",
-			},
-		},
-	},
-	"featureB": {
-		Id:   "featureB",
-		Name: "featureB",
-	},
-	"featureC": {
-		Id:   "featureC",
-		Name: "featureC",
-		Prerequisites: []*featureproto.Prerequisite{
-			{
-				FeatureId: "featureL",
-			},
-		},
-	},
-	"featureD": {
-		Id:   "featureD",
-		Name: "featureD",
-	},
-	"featureE": {
-		Id:   "featureE",
-		Name: "featureE",
-		Prerequisites: []*featureproto.Prerequisite{
-			{
-				FeatureId: "featureG",
-			},
-		},
-	},
-	"featureF": {
-		Id:   "featureF",
-		Name: "featureF",
-	},
-	"featureG": {
-		Id:   "featureG",
-		Name: "featureG",
-		Prerequisites: []*featureproto.Prerequisite{
-			{
-				FeatureId: "featureH",
-			},
-		},
-	},
-	"featureH": {
-		Id:   "featureH",
-		Name: "featureH",
-		Prerequisites: []*featureproto.Prerequisite{
-			{
-				FeatureId: "featureI",
-			},
-			{
-				FeatureId: "featureJ",
-			},
-		},
-	},
-	"featureI": {
-		Id:   "featureI",
-		Name: "featureI",
-		Prerequisites: []*featureproto.Prerequisite{
-			{
-				FeatureId: "featureK",
-			},
-		},
-	},
-	"featureJ": {
-		Id:   "featureJ",
-		Name: "featureJ",
-	},
-	"featureK": {
-		Id:   "featureK",
-		Name: "featureK",
-	},
-	"featureL": {
-		Id:   "featureL",
-		Name: "featureL",
-		Prerequisites: []*featureproto.Prerequisite{
-			{
-				FeatureId: "featureM",
-			},
-			{
-				FeatureId: "featureN",
-			},
-		},
-	},
-	"featureM": {
-		Id:   "featureM",
-		Name: "featureM",
-	},
-	"featureN": {
-		Id:   "featureN",
-		Name: "featureN",
-	},
-}
-
-func TestGetPrerequisiteDownwards(t *testing.T) {
-	t.Parallel()
-	mockController := gomock.NewController(t)
-	defer mockController.Finish()
-
-	patterns := []struct {
-		desc        string
-		target      []*featureproto.Feature
-		expected    []*featureproto.Feature
-		expectedErr error
-	}{
-		{
-			desc: "success: No prerequisites",
-			target: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureB"],
-				allFeaturesForPrerequisiteTest["featureD"],
-			},
-			expected: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureB"],
-				allFeaturesForPrerequisiteTest["featureD"],
-			},
-			expectedErr: nil,
-		},
-		{
-			desc: "success: Get prerequisites pattern1",
-			target: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureA"],
-			},
-			expected: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureA"],
-				allFeaturesForPrerequisiteTest["featureE"],
-				allFeaturesForPrerequisiteTest["featureF"],
-				allFeaturesForPrerequisiteTest["featureG"],
-				allFeaturesForPrerequisiteTest["featureH"],
-				allFeaturesForPrerequisiteTest["featureI"],
-				allFeaturesForPrerequisiteTest["featureJ"],
-				allFeaturesForPrerequisiteTest["featureK"],
-			},
-			expectedErr: nil,
-		},
-		{
-			desc: "success: Get prerequisites pattern2",
-			target: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureC"],
-				allFeaturesForPrerequisiteTest["featureD"],
-			},
-			expected: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureC"],
-				allFeaturesForPrerequisiteTest["featureD"],
-				allFeaturesForPrerequisiteTest["featureL"],
-				allFeaturesForPrerequisiteTest["featureM"],
-				allFeaturesForPrerequisiteTest["featureN"],
-			},
-			expectedErr: nil,
-		},
-		{
-			desc: "success: Get prerequisites pattern3",
-			target: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureD"],
-				allFeaturesForPrerequisiteTest["featureH"],
-			},
-			expected: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureD"],
-				allFeaturesForPrerequisiteTest["featureH"],
-				allFeaturesForPrerequisiteTest["featureI"],
-				allFeaturesForPrerequisiteTest["featureJ"],
-				allFeaturesForPrerequisiteTest["featureK"],
-			},
-			expectedErr: nil,
-		},
-	}
-	allFeatures := make([]*featureproto.Feature, 0, len(allFeaturesForPrerequisiteTest))
-	for _, v := range allFeaturesForPrerequisiteTest {
-		allFeatures = append(allFeatures, v)
-	}
-	for _, p := range patterns {
-		t.Run(p.desc, func(t *testing.T) {
-			gs := newGrpcGatewayServiceWithMock(t, mockController)
-			actual, err := gs.getPrerequisiteDownwards(p.target, allFeatures)
-			assert.ElementsMatch(t, p.expected, actual)
-			assert.Equal(t, p.expectedErr, err)
-		})
-	}
-}
-
-func TestGetPrerequisiteUpwards(t *testing.T) {
-	t.Parallel()
-	mockController := gomock.NewController(t)
-	defer mockController.Finish()
-
-	patterns := []struct {
-		desc        string
-		target      []*featureproto.Feature
-		expected    []*featureproto.Feature
-		expectedErr error
-	}{
-		{
-			desc: "success: No prerequisites",
-			target: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureA"],
-				allFeaturesForPrerequisiteTest["featureB"],
-				allFeaturesForPrerequisiteTest["featureC"],
-				allFeaturesForPrerequisiteTest["featureD"],
-			},
-			expected: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureA"],
-				allFeaturesForPrerequisiteTest["featureB"],
-				allFeaturesForPrerequisiteTest["featureC"],
-				allFeaturesForPrerequisiteTest["featureD"],
-			},
-			expectedErr: nil,
-		},
-		{
-			desc: "success: Get prerequisites pattern1",
-			target: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureF"],
-			},
-			expected: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureA"],
-				allFeaturesForPrerequisiteTest["featureF"],
-			},
-			expectedErr: nil,
-		},
-		{
-			desc: "success: Get prerequisites pattern2",
-			target: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureK"],
-				allFeaturesForPrerequisiteTest["featureE"],
-			},
-			expected: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureA"],
-				allFeaturesForPrerequisiteTest["featureE"],
-				allFeaturesForPrerequisiteTest["featureG"],
-				allFeaturesForPrerequisiteTest["featureH"],
-				allFeaturesForPrerequisiteTest["featureI"],
-				allFeaturesForPrerequisiteTest["featureK"],
-			},
-			expectedErr: nil,
-		},
-		{
-			desc: "success: Get prerequisites pattern3",
-			target: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureM"],
-				allFeaturesForPrerequisiteTest["featureN"],
-			},
-			expected: []*featureproto.Feature{
-				allFeaturesForPrerequisiteTest["featureC"],
-				allFeaturesForPrerequisiteTest["featureL"],
-				allFeaturesForPrerequisiteTest["featureM"],
-				allFeaturesForPrerequisiteTest["featureN"],
-			},
-			expectedErr: nil,
-		},
-	}
-	allFeatures := make([]*featureproto.Feature, 0, len(allFeaturesForPrerequisiteTest))
-	for _, v := range allFeaturesForPrerequisiteTest {
-		allFeatures = append(allFeatures, v)
-	}
-	for _, p := range patterns {
-		t.Run(p.desc, func(t *testing.T) {
-			gs := newGrpcGatewayServiceWithMock(t, mockController)
-			featuresHavePrerequisite := gs.getFeaturesHavePrerequisite(allFeatures)
-			actual, err := gs.getPrerequisiteUpwards(p.target, featuresHavePrerequisite)
-			assert.ElementsMatch(t, p.expected, actual)
-			assert.Equal(t, p.expectedErr, err)
-		})
-	}
-}
-
 func newUUID(t *testing.T) string {
 	t.Helper()
 	id, err := uuid.NewUUID()
